NEST code generation tests refactored
